Adds the JSF and EAR projects
This commit adds the JSF project with a web application that allows the
management of owners (by the administrators) and pets (by the owners).
The JSF project includes the functional tests for the owners management
page.

The EAR project only contains a POM that creates an EAR distribution of
the system.