Updates the project for the 16/17 course
The README.md has been update to reference the new course and the new
Eclipse version. In addition, the general pom.xml file has been updated
to include the newer versions of the dependencies, and the ear/pom.xml
has been modified to include more project information in the generated
EAR.

On the other hand, the testing mysql-ds.xml have been removed and the
standalone.xml testing files have been updated to include the datasource
data.