Improves the IsEqualToEntity abstract matcher
The IsEqualToEntity matcher (previously, IsEqualsToEntity) has been
improved to allow property comparison by functional reference to the
getter methods of the properties to compare. Specifically, the
"checkAttribute" methods added allow providing a Function<T, R> as a
getter, where T is the type of the entity to compare and R is the type
returned by the getter method.