Allows both application execution and deployment with Maven
The previous Maven configuration included a WildFly Maven plugin
configuration in the H2 and MySQL profiles that allowed the application
execution directly from Maven (with the mvn wildfly:start wildfly:deploy
command). As this configuration included a custom WildFly deployment
configuration (that modified the standalone.xml of the WildFly where the
application will be deployed), the mvn wildfly:deploy couldn't be used
to deploy the application in a local WildFly instance.

This commit adds a new profile (wildfly-mysql-run) that can be used to
execute the application directly from Maven (with the mvn install
command). The custom WildFly plugin configuration has been removed from
the other profiles, allowing users to normally deploy the application in
a running WildFly instances.

Instructions on how to use the profiles for application execution and
deployment can be found in the README.md file.