Adds Bootstrap as a Maven dependency
The Bootstrap libraries (CSS and JavaScript libraries) were inserted in
the web pages (JSF module) through a CDN. Now, these libraries are added
as Maven dependencies, using the Webjars project packages.

In addition, the link and script tags used in the JSF template have been
replaced by the h:outputStylesheet and h:outputScript tags from JSF, as
this tags make consistent the use of relative paths though the different
pages of the application.

Finally, a minor error has been solved in the pets.xhtml file, as the
placeholder text for the birth field was assigned to the animal type
field.